fix: argument bug fix

+++ program/Sender.py
@@ -25,7 +25,7 @@ class Sender(MDC):
         self._frame = None
         self._shape = (TARGET_HEIGHT, TARGET_WIDTH, TARGET_DEPTH)
         self._shared_memory_name = "jetson"
-        self._memory = posix_ipc.SharedMemory(self._shared_memory_name, flags=posix_ipc.O_CREAT, mode=0o777, size=np.prod(self._shape) * np.dtype(np.uint8).itemsize)
+        self._memory = posix_ipc.SharedMemory(self._shared_memory_name, flags=posix_ipc.O_CREAT, mode=0o777, size=int(np.prod(self._shape) * np.dtype(np.uint8).itemsize))
         self._map_file = mmap.mmap(self._memory.fd, self._memory.size)
         posix_ipc.close_fd(self._memory.fd)
